+++ src/helix-ui/elements/hx-select/index.js
@@ -1,6 +1,7 @@
-import { HXElement } from './HXElement';
-import shadowMarkup from './HXSelectElement.html';
-import shadowStyles from './HXSelectElement.less';
+import { HXElement } from '../HXElement.js';
+
+import shadowMarkup from './_shadow.html';
+import shadowStyles from './_shadow.less';
 
 /**
  * Applies Shadow DOM to the `<hx-select>` facade element.
