+++ src/helix-ui/elements/index.js
@@ -30,8 +30,8 @@ export { HXRadioSetElement } from './HXRadioSetElement';
 export { HXRevealElement } from './HXRevealElement';
 export { HXSearchAssistanceElement } from './hx-search-assistance/index.js';
 export { HXSearchElement } from './HXSearchElement';
-export { HXSelectControlElement } from './HXSelectControlElement';
-export { HXSelectElement } from './HXSelectElement';
+export { HXSelectControlElement } from './hx-select-control/index.js';
+export { HXSelectElement } from './hx-select/index.js';
 export { HXTabElement } from './hx-tab/index.js';
 export { HXTabcontentElement } from './hx-tabcontent/index.js';
 export { HXTablistElement } from './hx-tablist/index.js';
